Generate POJOs with non-nullable attributes with jOOQ
By default everything is nullable in the POJOs,
even if this is not the case in DB.
This would mean we have to handle the null cases somehow (eg. !!)
even when there is no need to do so.
With these flags, jOOQ will set the POJO attributes as non-nullable
if they are such in DB.